requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
openai==0.27.8
python-dotenv==1.0.0
Werkzeug==2.3.6
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
nltk==3.8.1
konlpy==0.6.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-caching==2.1.0
itsdangerous==2.1.2
markupsafe==2.1.3
jinja2==3.1.2
click==8.1.7
blinker==1.6.2
email-validator==2.0.0
wtforms==3.0.1
